# Scope requests, psutils, thop imports