--- a/ultralytics/data/loaders.py
+++ b/ultralytics/data/loaders.py
@@ -4,10 +4,10 @@ import glob
 import math
 import os
 import time
+import urllib
 from dataclasses import dataclass
 from pathlib import Path
 from threading import Thread
-from urllib.parse import urlparse
 
 import cv2
 import numpy as np
@@ -110,7 +110,7 @@ class LoadStreams:
         for i, s in enumerate(sources):  # index, source
             # Start thread to read frames from video stream
             st = f"{i + 1}/{n}: {s}... "
-            if urlparse(s).hostname in {"www.youtube.com", "youtube.com", "youtu.be"}:  # if source is YouTube video
+            if urllib.parse.urlparse(s).hostname in {"www.youtube.com", "youtube.com", "youtu.be"}:  # YouTube video
                 # YouTube format i.e. 'https://www.youtube.com/watch?v=Jsn8D3aC840' or 'https://youtu.be/Jsn8D3aC840'
                 s = get_best_youtube_url(s)
             s = eval(s) if s.isnumeric() else s  # i.e. s = '0' local webcam
@@ -585,12 +585,10 @@ class LoadTensor:
 
 def autocast_list(source):
     """Merges a list of sources into a list of numpy arrays or PIL images for Ultralytics prediction."""
-    import requests  # slow import
-
     files = []
     for im in source:
         if isinstance(im, (str, Path)):  # filename or uri
-            files.append(Image.open(requests.get(im, stream=True).raw if str(im).startswith("http") else im))
+            files.append(Image.open(urllib.request.urlopen(im) if str(im).startswith("http") else im))
         elif isinstance(im, (Image.Image, np.ndarray)):  # PIL or np Image
             files.append(im)
         else: