--- a/ultralytics/utils/dist.py
+++ b/ultralytics/utils/dist.py
@@ -2,7 +2,6 @@
 
 import os
 import shutil
-import socket
 import sys
 import tempfile
 
@@ -20,6 +19,8 @@ def find_free_network_port() -> int:
     Returns:
         (int): The available network port number.
     """
+    import socket
+
     with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
         s.bind(("127.0.0.1", 0))
         return s.getsockname()[1]  # port