**main**
`ultralytics: 3.7854s`

**PR**
`ultralytics: 3.0872s`

**Profiling imports**
```python
import ultralytics
import sys
modules  = set([m.split(".")[0] for m in sys.modules.keys() if not m.startswith('_')])  # all modules loaded by import
import subprocess

timings = []

def measure_import_time(module):
    code = f"""
import time
start = time.time()
import {module}
print(time.time() - start)
"""
    result = subprocess.run(["python", "-c", code], capture_output=True, text=True)
    try:
        return float(result.stdout.strip())
    except ValueError:
        return None

for mod in modules:
    t = measure_import_time(mod)
    if t is not None:
        timings.append((mod, t))

timings.sort(key=lambda x: x[1], reverse=True)

for mod, t in timings:
    print(f"{mod}: {t:.4f}s")
```

**Slowest imports triggered by Ultralytics**
```bash
ultralytics: 3.7854s
thop: 2.0518s
torch: 1.9941s
matplotlib: 0.4844s
sympy: 0.2702s
cv2: 0.2015s
numpy: 0.1794s
requests: 0.1188s
setuptools: 0.1175s
distutils: 0.1104s
pyparsing: 0.0864s
urllib3: 0.0664s
asyncio: 0.0350s
tqdm: 0.0227s
mpmath: 0.0214s
chardet: 0.0210s
pynvml: 0.0209s
psutil: 0.0196s
charset_normalizer: 0.0179s
pdb: 0.0170s
unittest: 0.0164s
yaml: 0.0161s
simplejson: 0.0131s
typing_extensions: 0.0120s
socks: 0.0113s
certifi: 0.0113s
pprint: 0.0108s
```

`sympy` import was due to `torch.utils.checkpoint`.

<!--
Thank you 🙏 for your contribution to [Ultralytics](https://www.ultralytics.com/) 🚀! Your effort in enhancing our repositories is greatly appreciated. To streamline the process and assist us in integrating your Pull Request (PR) effectively, please follow these steps:

1. Check for Existing Contributions: Before submitting, kindly explore existing PRs to ensure your contribution is unique and complementary.
2. Link Related Issues: If your PR addresses an open issue, please link it in your submission. This helps us better understand the context and impact of your contribution.
3. Elaborate Your Changes: Clearly articulate the purpose of your PR. Whether it's a bug fix or a new feature, a detailed description aids in a smoother integration process.
4. Ultralytics Contributor License Agreement (CLA): To uphold the quality and integrity of our project, we require all contributors to sign the CLA. Please confirm your agreement by commenting below:

    I have read the CLA Document and I sign the CLA

For more detailed guidance and best practices on contributing, refer to our ✅ [Contributing Guide](https://docs.ultralytics.com/help/contributing/). Your adherence to these guidelines ensures a faster and more effective review process.
-->


## 🛠️ PR Summary

<sub>Made with ❤️ by [Ultralytics Actions](https://github.com/ultralytics/actions)<sub>

### 🌟 Summary
This PR optimizes import statements across the Ultralytics codebase, moving slow or optional imports inside functions to speed up initial loading and reduce unnecessary dependencies. 🚀

### 📊 Key Changes
- Moved slow or optional imports (like `requests`, `psutil`, `hashlib`, `socket`, `thop`, and `uuid`) inside the functions where they're used, instead of importing them at the top of files.
- Updated code to use dynamic imports (e.g., `__import__("psutil")`) for certain modules.
- Ensured that optional dependencies (like `thop` for FLOPs profiling) are only imported when needed, with fallbacks if not installed.
- Cleaned up unused imports throughout the code.

### 🎯 Purpose & Impact
- **Faster Startup:** Reduces the time it takes to start Ultralytics tools by avoiding slow imports unless they're actually needed. ⚡
- **Lower Memory Usage:** Minimizes memory footprint by only loading modules when required.
- **Better Compatibility:** Prevents errors for users who don't have optional dependencies installed, making the codebase more robust and user-friendly.
- **Cleaner Code:** Improves code maintainability by removing unnecessary imports and clarifying where dependencies are needed.

Overall, these changes make Ultralytics tools more efficient and reliable for all users! 🛠️✨