@Y-T-G 2025-05-02
@glenn-jocher Importing Ultralytics even after the slow imports are already loaded (and hence wouldn't be loaded again) still takes 800+ms:

```python
In [1]: import torch, cv2, torchvision, numpy, matplotlib
%time
In [2]: %time import ultralytics
CPU times: user 829 ms, sys: 48 ms, total: 877 ms
Wall time: 878 ms
```

Probably need to find the cause of this.