@glenn-jocher 2025-05-02
I get these times for imports (ms) on M4 Pro

```
Import Speed Benchmarks:
 Package  Time (s)  1-sigma
requests    26.345    0.357
 urllib3     7.451    0.184
  urllib     0.000    0.000
```