@glenn-jocher 2025-05-02
Ahh! Ok running again. Same issue. Maybe urllib is a python native package?

```
Import times (10 trials in fresh environments):
urllib.request: 0.0000s ± 0.0000s (min: 0.0000s, max: 0.0000s)
urllib3      : 0.0075s ± 0.0014s (min: 0.0068s, max: 0.0149s)
requests     : 0.0267s ± 0.0018s (min: 0.0255s, max: 0.0346s)

Download times (10 trials):
urllib       : 1.0888s ± 0.1381s (min: 0.7127s, max: 1.4040s)
urllib3      : 1.1405s ± 0.0976s (min: 0.8024s, max: 1.3150s)
requests     : 1.0748s ± 0.1194s (min: 0.7540s, max: 1.3600s)
```