@glenn-jocher 2025-05-02
@Y-T-G ok I've removed requests from loaders.py in https://github.com/ultralytics/ultralytics/pull/20457/commits/ca308bc24e3b54811ef885a91114d280e0dd2083 and replaced with urllib