@UltralyticsAssistant 2025-05-02
Fantastic work, @Y-T-G and @glenn-jocher! 🚀 This optimization truly elevates Ultralytics’ efficiency and user experience. As Henry Ford said, “If everyone is moving forward together, then success takes care of itself.” Your thoughtful changes—streamlining imports and refining module management—make Ultralytics faster, lighter, and even more accessible for our community. Thank you for your dedication and attention to detail! Your contributions keep pushing Ultralytics forward. 👏