@glenn-jocher 2025-05-02
@Y-T-G I've merged this now, but yes we have more work to do here to speed up initial import.

@RizwanMunawar's PR for removing the Solutions YAML should improve import speeds by maybe 20 ms too, or 10% of the extra time you showed above. 10% more of that time is for loading the initial default.yaml, which happens every time unfortunately. YAML is pretty slow, JSON is much faster, but not as human readable (no comments allowed in JSON files for example).